refactor: renamed files and directories for clarity Refactor project structure to improve organization and naming conventions (#20)
