+++ app/services/trajectory_latest_service.py
