+++ app/services/trajectory_service.py
